Move generateETag and isFresh to src/utils
